This is a collection of classes that parallels the general (high-level) classes used in creating new language-service extensions.  These provide some analogous capabilities for creating things ("generated components") that are not extensions.  (The "new compiler" is an example of a generated component that is not an extension.)  The classes introduced here support the use of wizards to collect information and the use of that information to generate implementation classes based on provided templates, as is done for new extensions.  However, the classes introduced here to not add extension-related information to the plugin.xml file or perform other actions related to introducing a new extension.
Implementation note:  These classes are largely based on the corresponding classes for new extensions, wtih various amounts of unnecessary material trimmed away.  It may be possible to define some common superclasses that would support both new extensions and new non-extensions (or perhaps the classes for new extensions may be refactored into subclasses of the classes for new non-extensions).  Whether that is feasible and worthwhile I have left as an exercise for the future.