Changes related mainly to some subset of the following interrelated refactorings:
1)  Removing from generateCodeStubs methods the logic that computed the values of substitution parameters

2)  Localizing most of that logic, which is common across most wizards, in the collectCodeParms() method of CodeServiceWizard

3)  Adopting a standard set of (Java) fields to hold values obtained from the wizard fields for use as substitution parameters and in generating code stubs; also making use of these standard fields in the generateCodeStubs methods

4)  Where necessary, providing an wizard-specific collectCodeParms() method to collect wizard-specific parameters (where the wizard-specific method also calls the method on the supertype)

Relating to a separate issue, one other change that affects some wizards is to provide a method
    protected String[] getFilesThatCouldBeClobbered() ...
that returns an array of the names of the files that the wizard will generate if and when finished.  If this is just one implementation class (or even no implementation class) then the method is not needed (because that can be handled in codeServiceWizard).  If multiple files will be generated, then the specific wizard must implement this method and return the names of all files that will be generated.