Hoisted the "protected IProject fProject;", which can be standardized and used above the level of specific wizards.
Added a utility method
    public String getProjectSourceLocation()
to get that part of the path to a resource or file that falls between the end of the project path and the start of the package path.  Initially returns "src/".  Replaced uses of "src/" with calls to this method.


Also introduced several changes related to controlling the clobbering of existing files by newly generated ones:

- Added a method protected String[] getFilesThatCouldBeClobbered() to get the files that could be clobbered by a wizard (subtypes must override with a meaningful implementation).

- Added a method okToClobberFiles(..) that puts up a dialog to ask the user if it's ok to clobber files

- Adapted the method performFinish() to call okToClobberFiles (and return false if not)