At several points, added code to reloaded the extensions model using the IMP-based representation before accessing model details.
Refactored the signature of the version of addExstensions(..) that is not based on a wizard page to also take a project.

Significantly modified the internal logic of both versions of addExtensions(..) so as to always reload the extensions model using the IMP-base representation before accessing its details.

Added a method loadImpExtensionsModel(IPluginModel, IProject).

Modified logic in addRequiredPluginImports(..) to operate directly on the contents of the manifest file using String-based operations rather than trying to manipulate the model (which we cannot effectively save using the current pde.core mechanisms).

Commented out the two versions of addImports(..), the uses of which were redundant or unnecessary and were themselves commented out.

Modified saveAndRefresh() to save the extensions model rather than the bundlePluginModel (since the latter also saves imports from the model and we're currently not addressing imports through the model).

Cleaned up imports.  Minor reformatting and cleanup of comments.