Commented out some fields, related to the use of extension points, that were copied from ExtensionPointWizardPage but not needed here (where no extension point is involved).
Eliminated one unused constructor and modified the signatures of others to reflect the disuse of the pluginID field.

Changed the parameter and field name "pointID" to "compoentID" to reflect the fact that the id refers to a component still but not one that is defined in relation to an extension point.

Added a method to create an arbitrary text field for the page; this supports the method createAdditionalControls(..), which is the main way that wizard-specific fields will be created in the absence of an extension schema.

Added a method to get the project name from the project field on the page.

Modified one call of a subroutine to a method that was moved from ExtensionPointWizard to WizardUtilities.

May still need some cleaning and polishing.