Added a lot of functionality hoisted up from ExtensionPointWizardPage and GeneratedComponentWizardPage, where much of it was replicated.  Generalized original methods a little bit in some cases.
Two specific points of note:

Defined a field fComponentId to serve as a generalization of fExtPointID that could apply to  wizards and pages regardless of whether they represent extensions.  (It's generally helpful to have an internal name for what the specific service you're dealing with.)

Removed the semantically unclear getProject() function and replaced it with getProjectOfRecord() (which returns the  current value of the fProject field of the page), discoverSelectedProject() (which returns the currently selected project, regardless of the project named or set in the page), and getProjectBasedOnNameField() (which returns the project named by the value in the name field).