Renamed method createLanguageFieldForPlatformSchema(..) to createLanguageFieldForComponent(..).
Modified the renamed method createLanguageFieldForPlatformSchema(..) to create the language field in a disabled state (since for most wizards the language can be uniquely determined from the project--see Bug #303).

Commented out a number of calls to System.out.

Removed a large block of previously commented-out code (that had been moved to a utility package).