Modified so that NewNatureEnabler extends GeneratedComponentWizard rather than Wizard and NewNatureEnablerPage extends GeneratedComponentWizardPage rather than WizardPage.
Removed several methods and fields that were now available in the new supertypes.

Added method createAdditionalControls(Composite) to create a "class" field on the wizard page (which had previously been omitted) and to add a listener to the page's language field (based on which the default value of the "class" field is set).

Made modifications in several places related to removing assumptions about the names of the enabler class and package and determining those instead based on the value of the "class" field.  (This includes adding support for a "$ENABLER_CLASS_NAME$" substitution parameter.)