Changed the type of selection dialog used in the project browser to the new ListSelectionDialog (and provided that with a list of projects).
Added methods to support the dynamic setting and getting of selection filters and verifiers in the project-selection dialog and on the page itself.

Updated the diaogChanged() method to remove the hard-coding of some validation conditions for projects and replaced those with a call to a method that will validate projects according to dynamically given conditions.

Renamed method createProjectLabelText(..) (which did more than that) to createProjectField(..) (which better reflects the effect of the method).

Removed some previously commented-out methods and code sections.