Created another version of copyLiteralFile(..), one that provides a parameter for the name of the target file.
Created a method to createFileContentsFromTemplate(..) to return the contents of a template file as a String and made use of that in createFileFromTemplate and extendFileFromTemplate.  Adapted createFileFromTemplate(..) to use this.

Fixed extendFileFromTemplate to have correct error messages, to obtain file contents in a correct way, and to not append a closing brace on the extended file (file extensions must include the brace).

Added a check for a null TextEdit in formatJava Code(String).


?

?""""""""""""""''''