Principle change is to add a method:
    protected WizardDialog openClassDialog(
        String componentID, String interfaceQualName,
        String superClassName, Text text)

and supporting types, methods, and fields.  The principle purposes of these are to manage the require-bundle attribute in the plug-in manifest.mf file of the project under development (to assure that the bundle that supplies the "basedOn" class is accessible) and to assure that the class name field associated with this dialog has a (probably) meaningful value (that can be used by the class-creation wizard).  Motivated by solution to bug #391.